Complete monument card
