+++ src/header/styles.css
@@ -1,3 +1,5 @@
 .title {
   font-size: 3rem;
+  color: #eee;
+  text-align: center;
 }
