+++ src/index.ts
@@ -1,5 +1,6 @@
 import renderMainContainer from "./mainContainer/mainContainer";
 import monuments from "./monuments";
+
 import "./styles.css";
 
 renderMainContainer(monuments);
