+++ src/styles.css
@@ -16,7 +16,8 @@ body {
 }
 
 h1,
-h2 {
+h2,
+h3 {
   margin: 0;
 }
